pyproject.toml
[tool.poetry]
name = "ase-standup"
version = "0.1.0"
description = ""
authors = ["Terry Brooks - Learnosity  <Terry.Brooks@Learnosity.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.5"
psycopg2-binary = "^2.9.7"
gunicorn = "^21.2.0"
django-crispy-forms = "^2.0"
whitenoise = "^6.5.0"
pendulum = "^2.1.2"
python-dotenv = "^1.0.0"
crispy-bootstrap5 = "^0.7"
django-htmx = "^1.16.0"
black = "^23.7.0"
django-rest-framework = "^0.1.0"
django-filter = "^23.2"
pip = "^23.2.1"
install = "^1.3.5"
markdown = "^3.4.4"
loguru = "^0.7.1"
django-cors-headers = "^4.2.0"
django-password-reset = "^2.0"
pre-commit = "^3.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "django"
combine_as_imports = true
include_trailing_comma = true
line_length = 120

[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'


[tool.djlint]
profile="django"
ignore= 'H021,H017'
